requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
click==8.1.3
Flask==2.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.1.0
torch==1.9.0+cpu
torchvision==0.10.0+cpu
typing-extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
